+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This changelog contains a loose collection of changes in every release. I will a
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+* Support for wildcard-principal `"Principal": "*"` (#2)
+
 ## 1.1.0 - 2018-05-05
 
 ### Added